[Refactor] 파일 나누고 일 처리 메소드를 Teller 구조체로 옮김
